pull duplicated spec functions into a helper module
